tessla_server v0.0.1 TesslaServer.Computation.InputBuffer

Functionality to buffer inputs to a computations and get the next events to compute.

Types

event_map()

event_map :: %{optional(TesslaServer.GenComputation.id) => TesslaServer.Event.t}

t()

t :: %TesslaServer.Computation.InputBuffer{input_queues: queue}

Functions

add_event(buffer, event)

add_event(TesslaServer.Computation.InputBuffer.t, TesslaServer.Event.t) :: TesslaServer.Computation.InputBuffer.t

Adds an event to the appropriate queue in the buffer. If the buffer wasn’t initialized with the id of the event en error will be raised.

can_progress?(buffer)

can_progress?(TesslaServer.Computation.InputBuffer.t) :: boolean

Determines if the buffer holds enough events to progress. This means that there is an event present on each queues.

new(ids)

new([TesslaServer.GenComputation.id]) :: TesslaServer.Computation.InputBuffer.t

Sets up a new InputBuffer with empty queues for the given ids.

pop_head(buffer)

pop_head(TesslaServer.Computation.InputBuffer.t) :: {event_map | nil, Timex.Duration.t | nil, TesslaServer.Computation.InputBuffer.t}

Splits the buffer into an event map, containing all events of the queues at the minimal timestamp of all queues. If can_progress? returns true for the buffer, this function is guaranteed to return a non empty event_map, else the event_map will be nil and the buffer will be returned unmodified.