tessla_server v0.0.1 TesslaServer.Computation.Lifted.Min

Implements a Computation that compares two integer Streams and returns the smaller one

To do so the state.operands list has to be initialized with two integers representing the ids of the streams that should be compared

Functions

init_cache(state)

Callback implementation for TesslaServer.GenComputation.init_cache/1.

init_input_buffer(ids)

Callback implementation for TesslaServer.GenComputation.init_input_buffer/1.

output_event_type()

Callback implementation for TesslaServer.GenComputation.output_event_type/0.

process_event_map(event_map, timestamp, state)

Callback implementation for TesslaServer.GenComputation.process_event_map/3.

progress(state)

progress(TesslaServer.Computation.State.t) :: TesslaServer.Computation.State.t

propagate_output(event, state)

propagate_output(TesslaServer.Event.t, TesslaServer.Computation.State.t) :: :ok

propagate_progress(timestamp, state)

propagate_progress(Duration.t, TesslaServer.Computation.State.t) :: :ok

start(id, operands, options \\ %{})