tessla_server v0.0.1 TesslaServer.Event

Represents an Event on a Stream. Note that Event here is meant in a broader sense than the Tessla one, as it can represent an event on an eventstream, a change of a signal or the progress of a stream.

Types

event_type()

event_type :: :event | :change | :progress

t()

t :: %TesslaServer.Event{stream_id: TesslaServer.GenComputation.id, timestamp: Timex.Duration.t | :literal, type: event_type, value: any}