tessla_server v0.0.1 TesslaServer.EventStream

Represents a Stream, generated either from external Events or by computation through a Node Note that progressed_to can be bigger than the timestamp of the latest Event: If all Inputs of the Node generating this Stream have advanced to at least t but no new Event is generated (e.g. because the Node is a delay with a higher value) the Stream will be progressed to t but no Event with a timestamp t is present. progressed_to can also be smaller than timestamp of the latest function, when the events are generated by changing their timestamps (e.g. see TesslaServer.Node.Timing.Delay).

Types

stream_type :: :events | :signal

t :: %TesslaServer.EventStream{events: [TesslaServer.Event.t], id: integer | nil, progressed_to: timestamp | :literal, type: stream_type}

timestamp :: Timex.Types.timestamp

Functions

add_event(event_stream, event, progress \\ true)

Specs

add_event(nil | EventStream.t, TesslaServer.Event.t, boolean) ::
  {:ok, EventStream.t} |
  {:error, String.t}

Adds the Event to the EventStream.

iex> event = %Event{stream_id: :test, timestamp: {1000, 123456, 123456}, value: 1}
iex> stream = %EventStream{id: :test}
iex> {:ok, updated_stream} = EventStream.add_event(stream, event)
iex> hd(updated_stream.events)
%TesslaServer.Event{stream_id: :test, timestamp: {1000, 123456, 123456}, value: 1}

If the stream is nil it will create a new EventStream with the id of the event and progressed_to equal to the timestamp of the event

iex> event = %Event{stream_id: :test, timestamp: {1000, 123456, 123456}, value: 1}
iex> stream = nil
iex> {:ok, updated_stream} = EventStream.add_event(stream, event)
iex> updated_stream
%TesslaServer.EventStream{
  events: [%TesslaServer.Event{stream_id: :test, timestamp: {1000, 123456, 123456}, value: 1}],
  id: :test,
  progressed_to: {1000, 123456, 123456}
}

The timestamp of the Event has to be greater or equal to the progressed_to value of the EventStream.

iex> event = %Event{stream_id: :test, timestamp: {0, 1, 2}, value: 1}
iex> stream = %EventStream{id: :test, progressed_to: {1, 2, 3}}
iex> {:error, reason} = EventStream.add_event(stream, event)
iex> reason
"Event's timestamp smaller than stream progress"

The stream_id of the Event has to be the same as the id of the EventStream or else an error will be returned.

iex> event = %Event{stream_id: :wrong_id, timestamp: {1000, 123456, 123456}, value: 1}
iex> stream = %EventStream{id: :test}
iex> {:error, reason} = EventStream.add_event(stream, event)
iex> reason
"Event has different stream_id than stream"

This method will advance the progressed_to field to the timestamp of the Event.

event_at(stream, at)

Specs

event_at(EventStream.t, timestamp) ::
  TesslaServer.Event.t |
  nil

Returns the most recent Event.t that occured at or before at

Examples

iex> event0 = %Event{stream_id: :test, value: 0, timestamp: {0, 0, 0}}
iex> event1 = %Event{stream_id: :test, value: 1, timestamp: {1, 0, 0}}
iex> event2 = %Event{stream_id: :test, value: 2, timestamp: {2, 0, 0}}
iex> events = [event2, event1, event0]
iex> stream = %EventStream{id: :test, progressed_to: {2, 0, 0}, events: events}
iex> EventStream.event_at(stream, {1, 0, 0})
%Event{stream_id: :test, value: 1, timestamp: {1, 0, 0}}

iex> event2 = %Event{stream_id: :test, value: 2, timestamp: {2, 0, 0}}
iex> events = [event2]
iex> stream = %EventStream{id: :test, progressed_to: {2, 0, 0}, events: events}
iex> EventStream.event_at(stream, {1, 0, 0})
nil

events_in_timeslot(stream, from, to)

Specs

events_in_timeslot(EventStream.t | nil, timestamp, timestamp) :: [TesslaServer.Event.t]

Returns all event on the stream with a timestamp bigger than from and smaller or equal to to. To work, the stream has to be ordered by it’s timestamps, which all Stream.t structs should be always.

##Examples

iex> event0 = %Event{timestamp: {0, 0, 0}, stream_id: :test, value: 0}
iex> event1 = %Event{timestamp: {1, 0, 0}, stream_id: :test, value: 1}
iex> event2 = %Event{timestamp: {2, 0, 0}, stream_id: :test, value: 2}
iex> event3 = %Event{timestamp: {3, 0, 0}, stream_id: :test, value: 3}
iex> event4 = %Event{timestamp: {4, 0, 0}, stream_id: :test, value: 4}
iex> events = [event4, event3, event2, event1, event0]
iex> stream = %EventStream{id: :test, progressed_to: {4, 0, 0}, events: events}
iex> EventStream.events_in_timeslot(stream, {1, 0, 0}, {3,0,0})
[%Event{timestamp: {3, 0, 0}, stream_id: :test, value: 3},
%Event{timestamp: {2, 0, 0}, stream_id: :test, value: 2}]

progress(stream, timestamp)

Specs

progress(EventStream.t, timestamp) ::
  {:ok, EventStream.t} |
  {:error, String.t}

Progresses the EventStream to the given timestamp.

iex> timestamp = {1234, 123456, 123456}
iex> stream = %EventStream{}
iex> EventStream.progress(stream, timestamp)
{:ok, %EventStream{progressed_to: {1234, 123456, 123456}}}

The timestamp has to be bigger or equal to the streams progressed_to or else an error will be returned.

iex> timestamp = {1, 0, 0}
iex> stream = %EventStream{progressed_to: {2, 0, 0}}
iex> EventStream.progress(stream, timestamp)
{:error, "Timestamp smaller than progress of EventStream"}