tessla_server v0.0.1 TesslaServer.GenComputation behaviour

Base Module to work with Modules implementing the GenComputation behaviour. Note that Computations should implement specific GenServer calls, this cant be expressed right now by callbacks because the callbacks are allready defined in GenServer. Look at the methods below to see which messages are sent to a Computation.

Types

id()

id :: integer | nil

Functions

add_child(parent, child)

add_child(id, id) :: :ok

Adds a new child to the specified Computation

send_event(id, event)

send_event(id, TesslaServer.Event.t) :: :ok

Sends a new Event to the Computation that is registered with id to process it

stop(id)

stop(id) :: :ok

Stops a Computation and unregisters its id from gproc

subscribe_to_operands(id)

subscribe_to_operands(id) :: :ok

Subscribes the Computation specified by id to all Computation it is a descendant of.

Callbacks

init_cache(arg0)

init_cache(TesslaServer.Computation.State.t) :: TesslaServer.Computation.State.cache

init_input_buffer(list)

init_input_buffer([GenComputation.id]) :: TesslaServer.Computation.InputBuffer.t

output_event_type()

output_event_type :: TesslaServer.Event.event_type

process_event_map(arg0, arg1, arg2)

process_event_map(TesslaServer.Computation.InputBuffer.event_map, Duration.t, TesslaServer.Computation.State.t) ::
  {:ok, events, TesslaServer.Computation.State.cache} |
  {:wait, TesslaServer.Computation.State.cache} |
  {:progress, TesslaServer.Computation.State.cache}