tessla_server v0.0.1 TesslaServer.Node behaviour

Base Module to build new Nodes

When you want to implement a new Node you should call use TesslaServer.Node Furthermore you’d have to implement the prepare_values and process_values functions.

Types

name :: atom | String.t

on_process ::
  {:ok, :wait} |
  {:ok, TesslaServer.Event.t}

Functions

add_child(parent, child)

Specs

add_child(name, name) :: :ok

Adds a named child to the named Node

get_history(name)

Specs

get_history(name) :: TesslaServer.Node.State.t

Gets the State of the Node that is registered under name

get_latest_output(name)

Specs

get_latest_output(name) :: any

Returns the latest output of the named Node

send_event(name, event)

Specs

send_event(name, TesslaServer.Event.t) :: :ok

Sends a new Event to the Node that is registered with name to process it

stop(name)

Specs

stop(name) :: :ok

Stops a Node and unregisters it’s name from gproc

Callbacks

prepare_values(list)

Specs

prepare_values([{:state, TesslaServer.Node.State.t}]) :: {:ok, [TesslaServer.Event.t]}

process_values(arg0, list)

Specs

process_values(TesslaServer.Node.State.t, [TesslaServer.Event.t]) :: TesslaServer.Node.on_process

start(%{})

Specs

start(%{stream_name: atom | String.t}) ::
  atom |
  String.t

will_add_child(arg0, name)

Specs

will_add_child(TesslaServer.Node.State.t, name) :: boolean