tessla_server v0.0.1 TesslaServer.Node.Add

Implements a Node that adds two event streams

To do so the state.options object has to be initialized with the keys :summand1 and :summand2, which must be atoms representing the names of the event streams that should be summed.

Functions

process(event, state)

Specs

process(TesslaServer.Event.t, TesslaServer.Node.State.t) :: TesslaServer.Node.on_process

Processes a given event by adding it’s value to the last value of the other event stream saved in state.history. To do so the keys :summand1 and :summand2 have to be present in state.options, and the value of one of them has to be the same as event.stream_name and the value of the other one has to match a key in state.history.inputs

Returns :wait if at least one event stream doesn’t have a value until now or {:ok, new_event} where the new event is computed by adding the two last events of the specified event stream.