tessla_server v0.0.1 TesslaServer.Node.History

Contains the Structure and Methods to work with the History of a TesslaServer.Node

Types

t :: %TesslaServer.Node.History{inputs: input_streams, output: TesslaServer.EventStream.t}

timestamp :: Timex.Types.timestamp

Functions

latest_event_of_input_at(history, name, timestamp)

Specs

latest_event_of_input_at(TesslaServer.Node.History.t, atom, timestamp) ::
  TesslaServer.Event.t |
  nil

Returns the latest Event.t of the input stream specified by name in history that has a timestamp smaller than at

latest_input_event_at(history, timestamp)

Specs

latest_input_event_at(TesslaServer.Node.History.t, timestamp) ::
  TesslaServer.Event.t |
  nil

Returns the latest Event.t that is saved on any input stream

latest_output(history)

Specs

latest_output(TesslaServer.Node.History.t) :: TesslaServer.Event.t

Returns the latest Event in the output of a History

processable_events(history)

Specs

processable_events(TesslaServer.Node.History.t) :: [TesslaServer.Event.t]

Returns all events on all inputs with a timestamp between progressed_to of the output and the minimum of the progressed_to of all inputs.

progress_output(history, timestamp)

Specs

progress_output(TesslaServer.Node.History.t, timestamp) :: TesslaServer.Node.History

Sets the progressed_to of the output of the history to the timestamp

replace_input_stream(history, stream)

Specs

replace_input_stream(TesslaServer.Node.History.t, TesslaServer.EventStream.t) :: {:ok, TesslaServer.Node.History.t}

Replaces the input EventStream with the same name as stream of the history. If the history doesn’t include an EventStream with the name an error will be returned.

Examples

iex> stream = %EventStream{name: :test, progressed_to: {1, 2, 3}}
iex> history = %History{inputs: %{test: stream}}
iex> new_stream = %EventStream{name: :test, progressed_to: {2, 3, 4}}
iex> History.replace_input_stream(history, new_stream)
{:ok, %History{inputs: %{test: %EventStream{name: :test, progressed_to: {2, 3, 4}}}}}

iex> history = %History{}
iex> new_stream = %EventStream{name: :test, progressed_to: {2, 3, 4}}
iex> History.replace_input_stream(history, new_stream)
{:ok, %History{inputs: %{test: %EventStream{name: :test, progressed_to: {2, 3, 4}}}}}

update_output(history, new_event)

Specs

update_output(TesslaServer.Node.History.t, TesslaServer.Event.t) :: TesslaServer.Node.History.t

Updates the given history to prepend the given new_event to the output stream

Returns the new history