tessla_server v0.0.1 TesslaServer.Node.Lifted.Mul

Implements a Node that multiplies two event streams To do so the state.options object has to be initialized with the keys :operand1 and :operand2, which must be atoms representing the names of the event streams that should be multiplied.

Functions

prepare_values(state)

Callback implementation for TesslaServer.Node.prepare_values/1.

process_values(state, events)

Callback implementation for TesslaServer.Node.process_values/2.

start(args)

Callback implementation for TesslaServer.Node.start/1.

will_add_child(, )

Callback implementation for TesslaServer.Node.will_add_child/2.