tessla_server v0.0.1 TesslaServer.Node.Lifted.Or

Implements a Node that performs an or on two boolean streams

To do so the state.operands list has to be initialized with two integers representing the ids of the streams which should be the base of the computation

Functions

init_inputs(ids)

Callback implementation for TesslaServer.Node.init_inputs/1.

init_output(state)

Callback implementation for TesslaServer.Node.init_output/1.

perform_computation(timestamp, event_map, state)

Callback implementation for TesslaServer.Node.perform_computation/3.

prepare_events(at, state)

Callback implementation for TesslaServer.Node.prepare_events/2.

process_events(timestamp, event_map, state)

Callback implementation for TesslaServer.Node.process_events/3.

progress(state)

Specs

progress(TesslaServer.Node.State.t) :: TesslaServer.Node.State.t

start(id, operands, options \\ %{})

Callback implementation for TesslaServer.Node.start/3.