tessla_server v0.0.1 TesslaServer.Output

Used to log outputs. Has to be initialized at startup to hold the ids and names of streams that should be logged.

Functions

log_new_event(id, event)

log_new_event(TesslaServer.GenComputation.id, TesslaServer.Event.t) :: :ok

start(map \\ %{})

start(%{optional(TesslaServer.GenComputation.id) => String.t}) :: {:ok, pid}

stop()

stop :: {:ok, pid}