tessla_server v0.0.1 TesslaServer.Node behaviour

Base Module to build new Nodes

When you want to implement a new Node you should call use TesslaServer.Node Furthermore you’d have to implement the prepare_values and process_values functions.

Types

id :: integer

on_process ::
  {:ok, :wait} |
  {:ok, TesslaServer.Event.t}

Functions

add_child(parent, child)

Specs

add_child(id, id) :: :ok

Adds a new child to the specifiedied Node

get_history(id)

Specs

get_history(id) :: TesslaServer.Node.State.t

Gets the State.t of the Node that is registered under id

get_latest_output(id)

Specs

get_latest_output(id) :: any

Returns the latest output of the specified Node

log_new_outputs(id, events)

Specs

log_new_outputs(String.t, [TesslaServer.Event.t]) :: nil

send_event(id, event)

Specs

send_event(id, TesslaServer.Event.t) :: :ok

Sends a new Event to the Node that is registered with id to process it

stop(id)

Specs

stop(id) :: :ok

Stops a Node and unregisters it’s id from gproc

subscribe_to_operands(id)

Specs

subscribe_to_operands(id) :: :ok

Subscribes the Node specified by id to all Nodes it is a descendant of.

update_input_stream(id, stream)

Specs

update_input_stream(id, TesslaServer.EventStream.t) :: :ok

Sends the Node specified by id an EventStream.t so that it can update it’s inputs.

Callbacks

init_inputs(list)

Specs

init_inputs([id]) :: %{id => TesslaServer.EventStream.t}

init_output(arg0)

Specs

init_output(TesslaServer.Node.State.t) :: TesslaServer.EventStream.t

perform_computation(timestamp, event_map, arg2)

Specs

perform_computation(timestamp, event_map, TesslaServer.Node.State.t) :: computed_event

prepare_events(timestamp, arg1)

Specs

prepare_events(timestamp, TesslaServer.Node.State.t) :: event_map

process_events(timestamp, event_map, arg2)

Specs

process_events(timestamp, event_map, TesslaServer.Node.State.t) :: TesslaServer.Node.State.t

start(id, list, %{})

Specs

start(id, [id], %{}) :: id