tessla_server v0.0.1 TesslaServer.Node behaviour

Base Module to build new Nodes

When you want to implement a new Node you should call use TesslaServer.Node Furthermore you’d have to implement the prepare_values and process_values functions.

Types

name :: atom | String.t

on_process ::
  {:wait, TesslaServer.Node.State.t} |
  {:ok, %{event: TesslaServer.Event.t, state: TesslaServer.Node.State.t}}

Functions

get_history(name)

Specs

get_history(name) :: TesslaServer.Node.State.t

Gets the State of the Node that is registered under name

get_latest_output(name)

Specs

get_latest_output(name) :: any

send_event(name, event)

Specs

send_event(name, TesslaServer.Event.t) :: :ok

Sends a new Event to the Node that is registered with name to process it

Callbacks

prepare_values(list)

Specs

prepare_values([{:state, TesslaServer.Node.State.t}]) :: prepared_values

process_values(prepared_values)

Specs

process_values(prepared_values) :: TesslaServer.Node.on_process

start(%{})

Specs

start(%{stream_name: atom | String.t}) ::
  atom |
  String.t