tessla_server v0.0.1 TesslaServer.Node.Add

Implements a Node that adds two event streams

To do so the state.options object has to be initialized with the keys :summand1 and :summand2, which must be atoms representing the names of the event streams that should be summed.

Functions

prepare_values(state)

Specs

prepare_values(TesslaServer.Node.State.t) :: %{values: [TesslaServer.Event.t], state: TesslaServer.Node.State.t}

Callback implementation for TesslaServer.Node.prepare_values/1.

process_values(map)

Specs

process_values(%{values: [TesslaServer.Event.t], state: TesslaServer.Node.State.t}) :: TesslaServer.Node.on_process

Callback implementation for TesslaServer.Node.process_values/1.