tessla_server v0.0.1 TesslaServer.Node.History

Contains the Structure and Methods to work with the History of a TesslaServer.Node

Types

t :: %TesslaServer.Node.History{inputs: input_streams, output: event_stream}

Functions

get_latest_input(history)

Specs

get_latest_input(TesslaServer.Node.History.t) :: TesslaServer.Event.t

Returns the latest Event.t that is saved on any input stream It assumes that every Stream is ordered by time, meaning it only looks at the head of every Stream

get_latest_input_of_stream(history, name)

Specs

get_latest_input_of_stream(TesslaServer.Node.History.t, atom) ::
  TesslaServer.Event.t |
  nil

Returns the latest Event.t of the input stream specified by name in history

get_latest_output(history)

Specs

get_latest_output(TesslaServer.Node.History.t) :: TesslaServer.Event.t

Returns the latest Event in the output of a History

update_input(history, new_event)

Specs

update_input(TesslaServer.Node.History.t, TesslaServer.Event.t) :: TesslaServer.Node.History.t

Updates the given history to prepend the given event to the input stream specified by the new_event

Returns the new History

iex> history = %History{} iex> timestamp = Timex.Time.zero iex> event = %Event{stream_name: :test, value: :value, timestamp: timestamp} iex> updated = History.update_input history, event iex> updated.inputs[:test] |> hd %TesslaServer.Event{stream_name: :test, timestamp: {0, 0, 0}, value: :value}

update_output(history, new_event)

Specs

update_output(TesslaServer.Node.History.t, TesslaServer.Event.t) :: TesslaServer.Node.History.t

Updates the given history to prepend the given new_event to the output stream

Returns the new history